Split up some prep code, added util to read in ntuple variable names from a config
