+++ config/prepping/var_names_config.json
@@ -1,0 +1,84 @@
+{
+    "branches":
+    {
+        "nominal" : "nominal",
+        "sysUP" : "CategoryReduction_JET_Pileup_RhoTopology__1up",
+        "sysDOWN" : "CategoryReduction_JET_Pileup_RhoTopology__1down"
+    },
+    "observables":
+    {
+        "isSemiLep" : "semileptonicEvent",
+        "jet_pt" : "jet_pt",
+        "jet_eta" : "jet_eta",
+        "jet_phi" : "jet_phi",
+        "jet_e" : "jet_E",
+        "jet_m" : null,
+        "jet_n" : "jet_n",
+        "jet_isbtag" : "jet_isbtagged_DL1r_70",
+        "jet_partonLabel" : "jet_truthPartonLabel",
+        "jet_isTruth" : null,
+        "lep_pt" : "lep_pt",
+        "lep_eta" : "lep_eta",
+        "lep_phi" : "lep_phi",
+        "met_met" : "met_met",
+        "met_phi" : "met_phi",
+        "th_pt" : "MC_thad_afterFSR_pt",
+        "th_eta" : "MC_thad_afterFSR_eta",
+        "th_phi" : "MC_thad_afterFSR_phi",
+        "th_m" : "MC_thad_afterFSR_m",
+        "th_y" : "MC_thad_afterFSR_y",
+        "th_E" : "MC_thad_afterFSR_y",
+        "th_pout" : "MC_thad_Pout",
+        "tl_pt" : "MC_tlep_afterFSR_pt",
+        "tl_eta" : "MC_tlep_afterFSR_eta",
+        "tl_phi" : "MC_tlep_afterFSR_phi",
+        "tl_m" : "MC_tlep_afterFSR_m",
+        "tl_y" : "MC_tlep_afterFSR_y",
+        "tl_E" : "MC_tlep_afterFSR_y",
+        "tl_pout" : "MC_tlep_Pout",
+        "ttbar_pt" : "MC_ttbar_afterFSR_pt",
+        "ttbar_eta" : "MC_ttbar_afterFSR_eta",
+        "ttbar_phi" : "MC_ttbar_afterFSR_phi",
+        "ttbar_m" : "MC_ttbar_afterFSR_m",
+        "ttbar_y" : "MC_ttbar_afterFSR_y",
+        "ttbar_E" : "MC_ttbar_afterFSR_E",
+        "ttbar_dphi" : "MC_deltaPhi_tt",
+        "ttbar_Ht" : "MC_HT_tt",
+        "ttbar_chi" : "MC_chi_tt",
+        "ttbar_yboost" : "MC_y_boost",
+        "b_pt" : "MC_b_afterFSR_pt",
+        "b_eta" : "MC_b_afterFSR_eta",
+        "b_phi" : "MC_b_afterFSR_phi",
+        "b_m" : "MC_b_afterFSR_m",
+        "bbar_pt" : "MC_bbar_afterFSR_pt",
+        "bbar_eta" : "MC_bbar_afterFSR_eta",
+        "bbar_phi" : "MC_bbar_afterFSR_phi",
+        "bbar_m" : "MC_bbar_afterFSR_m",
+        "bh_pt" : "MC_b_from_thad_pt",
+        "bh_eta" : "MC_b_from_thad_eta",
+        "bh_phi" : "MC_b_from_thad_phi",
+        "bh_m" : "MC_b_from_thad_m",
+        "bl_pt" : "MC_b_from_tlep_pt",
+        "bl_eta" : "MC_b_from_tlep_eta",
+        "bl_phi" : "MC_b_from_tlep_phi",
+        "bl_m" : "MC_b_from_tlep_m",
+        "wh_pt" : "MC_W_from_thad_afterFSR_pt",
+        "wh_eta" : "MC_W_from_thad_afterFSR_eta",
+        "wh_phi" : "MC_W_from_thad_afterFSR_phi",
+        "wh_m" : "MC_W_from_thad_afterFSR_m",
+        "wl_pt" : "MC_W_from_tlep_afterFSR_pt",
+        "wl_eta" : "MC_W_from_tlep_afterFSR_eta",
+        "wl_phi" : "MC_W_from_tlep_afterFSR_phi",
+        "wl_m" : "MC_W_from_tlep_afterFSR_m",
+        "wh_decay1_pt" : "MC_Wdecay1_from_thad_pt",
+        "wh_decay1_eta" : "MC_Wdecay1_from_thad_eta",
+        "wh_decay1_phi" : "MC_Wdecay1_from_thad_phi",
+        "wh_decay1_m" : "MC_Wdecay1_from_thad_m",
+        "wh_decay1_pdgid" : "MC_Wdecay1_from_thad_pdgid",
+        "wh_decay2_pt" : "MC_Wdecay2_from_thad_pt",
+        "wh_decay2_eta" : "MC_Wdecay2_from_thad_eta",
+        "wh_decay2_phi" : "MC_Wdecay2_from_thad_phi",
+        "wh_decay2_m" : "MC_Wdecay2_from_thad_m",
+        "wh_decay2_pdgid" : "MC_Wdecay2_from_thad_pdgid"
+    }
+}
